Updated return datatype for BmiService

+++ src/BmiService.java
@@ -1,7 +1,6 @@
 public class BmiService {
-    public double calculate(double heightInMeters, int weightInKilograms) {
+    public int calculate(double heightInMeters, int weightInKilograms) {
         // вес в килограммах разделить на квадрат роста в метрах
-        return weightInKilograms / (heightInMeters*heightInMeters);
+        return (int) (weightInKilograms / (heightInMeters*heightInMeters));
     }
 }
-
